int -> double

+++ MDCT/GeneralMDCT.h
@@ -7,7 +7,7 @@
 #define GENERAL_MDCT_H
 
 int MDCT(const int M, 
-         const int* src, 
+         const double* src, 
          double* dst);
 
 int iMDCT(const int M,
